fix(console): Attempt to fix matched bracket theme in console
Still some work to be done, here. When the cursor is on a bracket, a blue border appears around the box containing the matching bracket. This is working as intended in the source editor, but the border is getting clipped in the console. I'm unsure whether this can be overcome with theming, but it's entirely possible that it could be.